Refactor: 동아리 등록 API성공(#21)
기존 테이블과 jpa코드 사이의 불일치해결

+++ src/main/java/com/partnerd/web/dto/clubDTO/ClubRegisterRequestDTO.java
@@ -12,8 +12,8 @@ public class ClubRegisterRequestDTO {
     private String name;
     private String intro;
     private List<ContactMethodDTO> contactMethod;
-    private int memberId;
-    private int categoryId;
+    private Long memberId;
+    private Long categoryId;
     private String activities;
     private MultipartFile bannerImage;  //이미지로직 아직 미구현
     private MultipartFile profileImage;
